# openapi.yaml
openapi: 3.1.0
info:
  title: Spotify Playlist Maker API
  description: API to search, create playlists, and add tracks on Spotify
  version: '1.0.0'
...
servers:
  - url: https://spotify-baikai.dylansouthard.info
paths:
...
  /tracks/search:
    get:
      tags: [Tracks]
      summary: Search for a track
      operationId: searchTracks
      parameters:
        - in: query
          name: query
          schema:
            type: string
          required: true
        - in: query
          name: limit
          schema:
            type: integer
          required: false
      responses:
        '200':
          description: Track results returned
...
  /tracks/top:
    get:
      tags: [Tracks]
      summary: Get user's top tracks
      operationId: getTopTracks
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          required: false
        - in: query
          name: time_range
          schema:
            type: string
            enum: [short_term, medium_term, long_term]
          required: false
      responses:
        '200':
          description: List of top tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    title:
                      type: string
                    artist:
                      type: string
  /refresh-token:
    post:
      tags: [Auth]
      summary: Refresh the user's access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: New access token returned
...
  /token-openai:
    post:
      tags: [Auth]
      summary: Exchanges authorization code for access token (ChatGPT plugin)
      operationId: tokenOpenAI
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                grant_type:
                  type: string
                code:
                  type: string
                redirect_uri:
                  type: string
                client_id:
                  type: string
                client_secret:
                  type: string
                refresh_token:
                  type: string
              required:
                - grant_type
      responses:
        '200':
          description: Access token response from Spotify
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
...
components:
  securitySchemes:
    OAuth2Auth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://spotify-baikai.dylansouthard.info/login-openai
          tokenUrl: https://spotify-baikai.dylansouthard.info/token-openai
          scopes:
            playlist-modify-private: Modify private playlists
            playlist-modify-public: Modify public playlists
            user-library-read: Read User library
            user-top-read: Read User Top entries
            user-read-playback-state: Get playback state
            user-read-recently-played: Read recently played tracks
            user-modify-playback-state: Modify playback state
            playlist-read-private: Find the reusable private playback buffer
  schemas:
...
    RefreshTokenRequest:
      type: object
      description: Request body to refresh Spotify access token
      properties:
        refresh_token:
          type: string
      required:
        - refresh_token
    TokenResponse:
      type: object
      description: Response returned after token exchange
      required:
        - access_token
        - token_type
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        token_type:
          type: string
security:
  - OAuth2Auth: []